requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
anyio==3.6.2
    # via starlette
asgiref==3.6.0
    # via django
bidict==0.22.1
    # via python-socketio
click==8.1.3
    # via uvicorn
django==4.2.1
    # via -r requirements.in
fastapi==0.95.1
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.4
    # via anyio
pydantic==1.10.7
    # via fastapi
python-dotenv==1.0.0
    # via -r requirements.in
python-engineio==4.4.1
    # via python-socketio
python-socketio==5.8.0
    # via -r requirements.in
sniffio==1.3.0
    # via anyio
sqlparse==0.4.4
    # via django
starlette==0.26.1
    # via fastapi
typing-extensions==4.5.0
    # via pydantic
uvicorn==0.22.0
    # via -r requirements.in
